Don't assume a script tag for jsonp transport
The JSONP transport [assumes there's a `<script>` tag in the
document][trouble]:

```js
var ref = document.getElementsByTagName('script')[0];
ref.parentNode.insertBefore(js, ref);
```

This can be a problem. For instance, if your script is loaded at the end
of the `<body>` tag, and you're rendering a react.js app on
`document.body` (like the examples [here][examples]), then `ref` will be
undefined.

This just adds the JSONP `<script>` to the `<head>`.

I was getting the same error message as #43, but I don't know if Twitter
auth is using JSONP, so I'm not sure if it's related.

[trouble]:
https://github.com/firebase/firebase-simple-login/blob/5f0fc6639fc0d4c669c208d3d9f10265afd5e961/js/src/simple-login/transports/JSONP.js#L100
[examples]:
http://facebook.github.io/react/blog/2013/11/05/thinking-in-react.html